@@ -1,40 +1,40 @@
 ﻿
-using System.Web.Optimization;
+//using System.Web.Optimization;
 
-namespace Swetugg.Web
-{
-	public class BundleConfig
-	{
-		// For more information on bundling, visit http://go.microsoft.com/fwlink/?LinkId=301862
-		public static void RegisterBundles(BundleCollection bundles)
-		{
-			bundles.Add(new ScriptBundle("~/bundles/jquery").Include(
-						"~/Scripts/jquery-{version}.js"));
-            bundles.Add(new ScriptBundle("~/bundles/globalize").Include(
-                        "~/Scripts/globalize/globalize.js",
-                        "~/Scripts/globalize/cultures/globalize.culture.sv-SE.js"));
+//namespace Swetugg.Web
+//{
+//	public class BundleConfig
+//	{
+//		// For more information on bundling, visit http://go.microsoft.com/fwlink/?LinkId=301862
+//		public static void RegisterBundles(BundleCollection bundles)
+//		{
+//			bundles.Add(new ScriptBundle("~/bundles/jquery").Include(
+//						"~/Scripts/jquery-{version}.js"));
+//            bundles.Add(new ScriptBundle("~/bundles/globalize").Include(
+//                        "~/Scripts/globalize/globalize.js",
+//                        "~/Scripts/globalize/cultures/globalize.culture.sv-SE.js"));
 
-			bundles.Add(new ScriptBundle("~/bundles/jqueryval").Include(
-						"~/Scripts/jquery.validate.js",
-                        "~/Scripts/jquery.validate.unobtrusive.js",
-                        "~/Scripts/jquery.validate.globalize.js"));
+//			bundles.Add(new ScriptBundle("~/bundles/jqueryval").Include(
+//						"~/Scripts/jquery.validate.js",
+//                        "~/Scripts/jquery.validate.unobtrusive.js",
+//                        "~/Scripts/jquery.validate.globalize.js"));
 
-			// Use the development version of Modernizr to develop with and learn from. Then, when you're
-			// ready for production, use the build tool at http://modernizr.com to pick only the tests you need.
-			bundles.Add(new ScriptBundle("~/bundles/modernizr").Include(
-						"~/Scripts/modernizr-*"));
+//			// Use the development version of Modernizr to develop with and learn from. Then, when you're
+//			// ready for production, use the build tool at http://modernizr.com to pick only the tests you need.
+//			bundles.Add(new ScriptBundle("~/bundles/modernizr").Include(
+//						"~/Scripts/modernizr-*"));
 
-			bundles.Add(new ScriptBundle("~/bundles/bootstrap").Include(
-					  "~/Scripts/bootstrap.js",
-					  "~/Scripts/respond.js",
-                      "~/Scripts/mvc-bootstrap.js"));
+//			bundles.Add(new ScriptBundle("~/bundles/bootstrap").Include(
+//					  "~/Scripts/bootstrap.js",
+//					  "~/Scripts/respond.js",
+//                      "~/Scripts/mvc-bootstrap.js"));
 
-            bundles.Add(new ScriptBundle("~/bundles/app").Include(
-                      "~/Scripts/app.js"));
+//            bundles.Add(new ScriptBundle("~/bundles/app").Include(
+//                      "~/Scripts/app.js"));
 
-			bundles.Add(new StyleBundle("~/Content/css").Include(
-					  "~/Content/bootstrap.css",
-					  "~/Content/site.css"));
-		}
-	}
-}
+//			bundles.Add(new StyleBundle("~/Content/css").Include(
+//					  "~/Content/bootstrap.css",
+//					  "~/Content/site.css"));
+//		}
+//	}
+//}
