@@ -1,148 +1,148 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Configuration;
-using System.Data.Entity;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Net;
-using System.Security.Claims;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Identity;
-using SendGrid;
-using SendGrid.Helpers.Mail;
-using Swetugg.Web.Models;
+﻿//using System;
+//using System.Collections.Generic;
+//using System.Configuration;
+//using Microsoft.EntityFrameworkCore;
+//using System.Linq;
+//using System.Linq.Expressions;
+//using System.Net;
+//using System.Security.Claims;
+//using System.Threading.Tasks;
+//using Microsoft.AspNetCore.Identity;
+//using SendGrid;
+//using SendGrid.Helpers.Mail;
+//using Swetugg.Web.Models;
 
-namespace Swetugg.Web
-{
-    public class EmailService : IIdentityMessageService
-    {
-        public Task SendAsync(IdentityMessage message)
-        {
-            return configSendGridasync(message);
-        }
+//namespace Swetugg.Web
+//{
+//    public class EmailService : IIdentityMessageService
+//    {
+//        public Task SendAsync(IdentityMessage message)
+//        {
+//            return configSendGridasync(message);
+//        }
 
-        private async Task configSendGridasync(IdentityMessage message)
-        {
-            var myMessage = new SendGridMessage();
-            myMessage.AddTo(message.Destination);
-            myMessage.From = new EmailAddress("info@swetugg.se", "Swetugg");
-            myMessage.Subject = message.Subject;
-            myMessage.PlainTextContent = message.Body;
-            myMessage.HtmlContent = message.Body;
+//        private async Task configSendGridasync(IdentityMessage message)
+//        {
+//            var myMessage = new SendGridMessage();
+//            myMessage.AddTo(message.Destination);
+//            myMessage.From = new EmailAddress("info@swetugg.se", "Swetugg");
+//            myMessage.Subject = message.Subject;
+//            myMessage.PlainTextContent = message.Body;
+//            myMessage.HtmlContent = message.Body;
 
-            var sendgridKey = ConfigurationManager.AppSettings["SendGrid_ApiKey"];
+//            var sendgridKey = ConfigurationManager.AppSettings["SendGrid_ApiKey"];
 
-            // Create a Web transport for sending email.
-            var transportWeb = new SendGridClient(sendgridKey);
+//            // Create a Web transport for sending email.
+//            var transportWeb = new SendGridClient(sendgridKey);
 
-            // Send the email.
-            try
-            {
-                await transportWeb.SendEmailAsync(myMessage);
-            }
-            catch (Exception)
-            {
-                //
-            }
-        }
+//            // Send the email.
+//            try
+//            {
+//                await transportWeb.SendEmailAsync(myMessage);
+//            }
+//            catch (Exception)
+//            {
+//                //
+//            }
+//        }
         
-    }
+//    }
 
-    public class SmsService : IIdentityMessageService
-    {
-        public Task SendAsync(IdentityMessage message)
-        {
-            // Plug in your SMS service here to send a text message.
-            return Task.FromResult(0);
-        }
-    }
+//    public class SmsService : IIdentityMessageService
+//    {
+//        public Task SendAsync(IdentityMessage message)
+//        {
+//            // Plug in your SMS service here to send a text message.
+//            return Task.FromResult(0);
+//        }
+//    }
 
-    public class ApplicationRoleManager : RoleManager<IdentityRole>
-    {
-        public ApplicationRoleManager(IRoleStore<IdentityRole, string> store) : base(store)
-        {
-        }
+//    public class ApplicationRoleManager : RoleManager<IdentityRole>
+//    {
+//        public ApplicationRoleManager(IRoleStore<IdentityRole, string> store) : base(store)
+//        {
+//        }
 
-        public static ApplicationRoleManager Create(IdentityFactoryOptions<ApplicationRoleManager> options, IOwinContext context)
-        {
-            var manager = new ApplicationRoleManager(new RoleStore<IdentityRole>(context.Get<ApplicationDbContext>()));
+//        public static ApplicationRoleManager Create(IdentityFactoryOptions<ApplicationRoleManager> options, IOwinContext context)
+//        {
+//            var manager = new ApplicationRoleManager(new RoleStore<IdentityRole>(context.Get<ApplicationDbContext>()));
             
-            return manager;
-        }
-    }
+//            return manager;
+//        }
+//    }
 
-    // Configure the application user manager used in this application. UserManager is defined in ASP.NET Identity and is used by the application.
-    public class ApplicationUserManager : UserManager<ApplicationUser>
-    {
-        public ApplicationUserManager(IUserStore<ApplicationUser> store)
-            : base(store)
-        {
-        }
+//    // Configure the application user manager used in this application. UserManager is defined in ASP.NET Identity and is used by the application.
+//    public class ApplicationUserManager : UserManager<ApplicationUser>
+//    {
+//        public ApplicationUserManager(IUserStore<ApplicationUser> store)
+//            : base(store)
+//        {
+//        }
 
-        public static ApplicationUserManager Create(IdentityFactoryOptions<ApplicationUserManager> options, IOwinContext context) 
-        {
-            var manager = new ApplicationUserManager(new UserStore<ApplicationUser>(context.Get<ApplicationDbContext>()));
-            // Configure validation logic for usernames
-            manager.UserValidator = new UserValidator<ApplicationUser>(manager)
-            {
-                AllowOnlyAlphanumericUserNames = false,
-                RequireUniqueEmail = true
-            };
+//        public static ApplicationUserManager Create(IdentityFactoryOptions<ApplicationUserManager> options, IOwinContext context) 
+//        {
+//            var manager = new ApplicationUserManager(new UserStore<ApplicationUser>(context.Get<ApplicationDbContext>()));
+//            // Configure validation logic for usernames
+//            manager.UserValidator = new UserValidator<ApplicationUser>(manager)
+//            {
+//                AllowOnlyAlphanumericUserNames = false,
+//                RequireUniqueEmail = true
+//            };
 
-            // Configure validation logic for passwords
-            manager.PasswordValidator = new PasswordValidator
-            {
-                RequiredLength = 6,
-                RequireNonLetterOrDigit = true,
-                RequireDigit = true,
-                RequireLowercase = true,
-                RequireUppercase = true,
-            };
+//            // Configure validation logic for passwords
+//            manager.PasswordValidator = new PasswordValidator
+//            {
+//                RequiredLength = 6,
+//                RequireNonLetterOrDigit = true,
+//                RequireDigit = true,
+//                RequireLowercase = true,
+//                RequireUppercase = true,
+//            };
 
-            // Configure user lockout defaults
-            manager.UserLockoutEnabledByDefault = true;
-            manager.DefaultAccountLockoutTimeSpan = TimeSpan.FromMinutes(5);
-            manager.MaxFailedAccessAttemptsBeforeLockout = 5;
+//            // Configure user lockout defaults
+//            manager.UserLockoutEnabledByDefault = true;
+//            manager.DefaultAccountLockoutTimeSpan = TimeSpan.FromMinutes(5);
+//            manager.MaxFailedAccessAttemptsBeforeLockout = 5;
 
-            // Register two factor authentication providers. This application uses Phone and Emails as a step of receiving a code for verifying the user
-            // You can write your own provider and plug it in here.
-            manager.RegisterTwoFactorProvider("Phone Code", new PhoneNumberTokenProvider<ApplicationUser>
-            {
-                MessageFormat = "Your security code is {0}"
-            });
-            manager.RegisterTwoFactorProvider("Email Code", new EmailTokenProvider<ApplicationUser>
-            {
-                Subject = "Security Code",
-                BodyFormat = "Your security code is {0}"
-            });
-            manager.EmailService = new EmailService();
-            manager.SmsService = new SmsService();
-            var dataProtectionProvider = options.DataProtectionProvider;
-            if (dataProtectionProvider != null)
-            {
-                manager.UserTokenProvider = 
-                    new DataProtectorTokenProvider<ApplicationUser>(dataProtectionProvider.Create("ASP.NET Identity"));
-            }
-            return manager;
-        }
-    }
+//            // Register two factor authentication providers. This application uses Phone and Emails as a step of receiving a code for verifying the user
+//            // You can write your own provider and plug it in here.
+//            manager.RegisterTwoFactorProvider("Phone Code", new PhoneNumberTokenProvider<ApplicationUser>
+//            {
+//                MessageFormat = "Your security code is {0}"
+//            });
+//            manager.RegisterTwoFactorProvider("Email Code", new EmailTokenProvider<ApplicationUser>
+//            {
+//                Subject = "Security Code",
+//                BodyFormat = "Your security code is {0}"
+//            });
+//            manager.EmailService = new EmailService();
+//            manager.SmsService = new SmsService();
+//            var dataProtectionProvider = options.DataProtectionProvider;
+//            if (dataProtectionProvider != null)
+//            {
+//                manager.UserTokenProvider = 
+//                    new DataProtectorTokenProvider<ApplicationUser>(dataProtectionProvider.Create("ASP.NET Identity"));
+//            }
+//            return manager;
+//        }
+//    }
 
-    // Configure the application sign-in manager which is used in this application.
-    public class ApplicationSignInManager : SignInManager<ApplicationUser, string>
-    {
-        public ApplicationSignInManager(ApplicationUserManager userManager, IAuthenticationManager authenticationManager)
-            : base(userManager, authenticationManager)
-        {
-        }
+//    // Configure the application sign-in manager which is used in this application.
+//    public class ApplicationSignInManager : SignInManager<ApplicationUser, string>
+//    {
+//        public ApplicationSignInManager(ApplicationUserManager userManager, IAuthenticationManager authenticationManager)
+//            : base(userManager, authenticationManager)
+//        {
+//        }
         
-        public override Task<ClaimsIdentity> CreateUserIdentityAsync(ApplicationUser user)
-        {
-            return user.GenerateUserIdentityAsync((ApplicationUserManager)UserManager);
-        }
+//        public override Task<ClaimsIdentity> CreateUserIdentityAsync(ApplicationUser user)
+//        {
+//            return user.GenerateUserIdentityAsync((ApplicationUserManager)UserManager);
+//        }
 
-        public static ApplicationSignInManager Create(IdentityFactoryOptions<ApplicationSignInManager> options, IOwinContext context)
-        {
-            return new ApplicationSignInManager(context.GetUserManager<ApplicationUserManager>(), context.Authentication);
-        }
-    }
-}
+//        public static ApplicationSignInManager Create(IdentityFactoryOptions<ApplicationSignInManager> options, IOwinContext context)
+//        {
+//            return new ApplicationSignInManager(context.GetUserManager<ApplicationUserManager>(), context.Authentication);
+//        }
+//    }
+//}
