@@ -1,21 +1,21 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿//using System;
+//using System.Collections.Generic;
+//using System.Linq;
 
 
-namespace Swetugg.Web
-{
-    public static class WebApiConfig
-    {
-        public static void Register(HttpConfiguration config)
-        {
-            config.MapHttpAttributeRoutes();
+//namespace Swetugg.Web
+//{
+//    public static class WebApiConfig
+//    {
+//        public static void Register(HttpConfiguration config)
+//        {
+//            config.MapHttpAttributeRoutes();
 
-            config.Routes.MapHttpRoute(
-                name: "DefaultApi",
-                routeTemplate: "api/{controller}/{id}",
-                defaults: new { id = RouteParameter.Optional }
-            );
-        }
-    }
-}
+//            config.Routes.MapHttpRoute(
+//                name: "DefaultApi",
+//                routeTemplate: "api/{controller}/{id}",
+//                defaults: new { id = RouteParameter.Optional }
+//            );
+//        }
+//    }
+//}
