@@ -1,4 +1,4 @@
-﻿@using Microsoft.AspNet.Identity.EntityFramework
+﻿@*@using Microsoft.AspNet.Identity.EntityFramework
 @using Swetugg.Web.Models
 @model List<ApplicationUser>
 @{
@@ -55,3 +55,4 @@
         </tbody>
     </table>
 </div>
+*@
