@@ -1,558 +1,558 @@
-﻿using System;
-using System.Configuration;
-using System.Globalization;
-using System.Linq;
-using System.Security.Claims;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
-using Swetugg.Web.Models;
+﻿//using System;
+//using System.Configuration;
+//using System.Globalization;
+//using System.Linq;
+//using System.Security.Claims;
+//using System.Threading.Tasks;
+//using Microsoft.AspNetCore.Authorization;
+//using Microsoft.AspNetCore.Identity;
+//using Microsoft.AspNetCore.Mvc;
+//using Microsoft.AspNetCore.Mvc.Rendering;
+//using Swetugg.Web.Models;
 
-namespace Swetugg.Web.Controllers
-{
-    [RequireHttps]
-    [Authorize]
-    public class AccountController : Microsoft.AspNetCore.Mvc.Controller
-    {
-        private ApplicationSignInManager _signInManager;
-        private ApplicationUserManager _userManager;
+//namespace Swetugg.Web.Controllers
+//{
+//    [RequireHttps]
+//    [Authorize]
+//    public class AccountController : Microsoft.AspNetCore.Mvc.Controller
+//    {
+//        private ApplicationSignInManager _signInManager;
+//        private ApplicationUserManager _userManager;
 
-        private bool _useSendGrid = false;
+//        private bool _useSendGrid = false;
 
-        public AccountController()
-        {
-            var str = ConfigurationManager.AppSettings["SendGrid_Messaging_Enabled"];
+//        public AccountController()
+//        {
+//            var str = ConfigurationManager.AppSettings["SendGrid_Messaging_Enabled"];
 
-            if (!string.IsNullOrEmpty(str) && str.Equals("true", StringComparison.InvariantCultureIgnoreCase))
-                _useSendGrid = true;
-        }
+//            if (!string.IsNullOrEmpty(str) && str.Equals("true", StringComparison.InvariantCultureIgnoreCase))
+//                _useSendGrid = true;
+//        }
 
-        /*public AccountController(ApplicationUserManager userManager, ApplicationSignInManager signInManager )
-        {
-            UserManager = userManager;
-            SignInManager = signInManager;
-        }*/
+//        /*public AccountController(ApplicationUserManager userManager, ApplicationSignInManager signInManager )
+//        {
+//            UserManager = userManager;
+//            SignInManager = signInManager;
+//        }*/
 
-        public ApplicationSignInManager SignInManager
-        {
-            get
-            {
-                return _signInManager ?? HttpContext.GetOwinContext().Get<ApplicationSignInManager>();
-            }
-            private set 
-            { 
-                _signInManager = value; 
-            }
-        }
+//        public ApplicationSignInManager SignInManager
+//        {
+//            get
+//            {
+//                return _signInManager ?? HttpContext.GetOwinContext().Get<ApplicationSignInManager>();
+//            }
+//            private set 
+//            { 
+//                _signInManager = value; 
+//            }
+//        }
 
-        public ApplicationUserManager UserManager
-        {
-            get
-            {
-                return _userManager ?? HttpContext.GetOwinContext().GetUserManager<ApplicationUserManager>();
-            }
-            private set
-            {
-                _userManager = value;
-            }
-        }
+//        public ApplicationUserManager UserManager
+//        {
+//            get
+//            {
+//                return _userManager ?? HttpContext.GetOwinContext().GetUserManager<ApplicationUserManager>();
+//            }
+//            private set
+//            {
+//                _userManager = value;
+//            }
+//        }
 
-        //
-        // GET: /Account/Login
-        [AllowAnonymous]
-        public Microsoft.AspNetCore.Mvc.ActionResult Login(string returnUrl)
-        {
-            ViewBag.ReturnUrl = returnUrl;
-            return View();
-        }
+//        //
+//        // GET: /Account/Login
+//        [AllowAnonymous]
+//        public Microsoft.AspNetCore.Mvc.ActionResult Login(string returnUrl)
+//        {
+//            ViewBag.ReturnUrl = returnUrl;
+//            return View();
+//        }
 
-        //
-        // POST: /Account/Login
-        [HttpPost]
-        [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> Login(LoginViewModel model, string returnUrl)
-        {
-            if (!ModelState.IsValid)
-            {
-                ViewBag.ReturnUrl = returnUrl;
-                return View(model);
-            }
+//        //
+//        // POST: /Account/Login
+//        [HttpPost]
+//        [AllowAnonymous]
+//        [ValidateAntiForgeryToken]
+//        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> Login(LoginViewModel model, string returnUrl)
+//        {
+//            if (!ModelState.IsValid)
+//            {
+//                ViewBag.ReturnUrl = returnUrl;
+//                return View(model);
+//            }
 
-            // This doesn't count login failures towards account lockout
-            // To enable password failures to trigger account lockout, change to shouldLockout: true
-            var result = await SignInManager.PasswordSignInAsync(model.Email, model.Password, model.RememberMe, shouldLockout: false);
-            switch (result)
-            {
-                case SignInStatus.Success:
-                    return RedirectToLocal(returnUrl);
-                case SignInStatus.LockedOut:
-                    return View("Lockout");
-                case SignInStatus.RequiresVerification:
-                    return RedirectToAction("SendCode", new { ReturnUrl = returnUrl, RememberMe = model.RememberMe });
-                case SignInStatus.Failure:
-                default:
-                    ModelState.AddModelError("", "Invalid login attempt.");
-                    ViewBag.ReturnUrl = returnUrl;
-                    return View(model);
-            }
-        }
+//            // This doesn't count login failures towards account lockout
+//            // To enable password failures to trigger account lockout, change to shouldLockout: true
+//            var result = await SignInManager.PasswordSignInAsync(model.Email, model.Password, model.RememberMe, shouldLockout: false);
+//            switch (result)
+//            {
+//                case SignInStatus.Success:
+//                    return RedirectToLocal(returnUrl);
+//                case SignInStatus.LockedOut:
+//                    return View("Lockout");
+//                case SignInStatus.RequiresVerification:
+//                    return RedirectToAction("SendCode", new { ReturnUrl = returnUrl, RememberMe = model.RememberMe });
+//                case SignInStatus.Failure:
+//                default:
+//                    ModelState.AddModelError("", "Invalid login attempt.");
+//                    ViewBag.ReturnUrl = returnUrl;
+//                    return View(model);
+//            }
+//        }
 
-        //
-        // GET: /Account/VerifyCode
-        [AllowAnonymous]
-        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> VerifyCode(string provider, string returnUrl, bool rememberMe)
-        {
-            // Require that the user has already logged in via username/password or external login
-            if (!await SignInManager.HasBeenVerifiedAsync())
-            {
-                return View("Error");
-            }
-            return View(new VerifyCodeViewModel { Provider = provider, ReturnUrl = returnUrl, RememberMe = rememberMe });
-        }
+//        //
+//        // GET: /Account/VerifyCode
+//        [AllowAnonymous]
+//        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> VerifyCode(string provider, string returnUrl, bool rememberMe)
+//        {
+//            // Require that the user has already logged in via username/password or external login
+//            if (!await SignInManager.HasBeenVerifiedAsync())
+//            {
+//                return View("Error");
+//            }
+//            return View(new VerifyCodeViewModel { Provider = provider, ReturnUrl = returnUrl, RememberMe = rememberMe });
+//        }
 
-        //
-        // POST: /Account/VerifyCode
-        [HttpPost]
-        [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> VerifyCode(VerifyCodeViewModel model)
-        {
-            if (!ModelState.IsValid)
-            {
-                return View(model);
-            }
+//        //
+//        // POST: /Account/VerifyCode
+//        [HttpPost]
+//        [AllowAnonymous]
+//        [ValidateAntiForgeryToken]
+//        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> VerifyCode(VerifyCodeViewModel model)
+//        {
+//            if (!ModelState.IsValid)
+//            {
+//                return View(model);
+//            }
 
-            // The following code protects for brute force attacks against the two factor codes. 
-            // If a user enters incorrect codes for a specified amount of time then the user account 
-            // will be locked out for a specified amount of time. 
-            // You can configure the account lockout settings in IdentityConfig
-            var result = await SignInManager.TwoFactorSignInAsync(model.Provider, model.Code, isPersistent:  model.RememberMe, rememberBrowser: model.RememberBrowser);
-            switch (result)
-            {
-                case SignInStatus.Success:
-                    return RedirectToLocal(model.ReturnUrl);
-                case SignInStatus.LockedOut:
-                    return View("Lockout");
-                case SignInStatus.Failure:
-                default:
-                    ModelState.AddModelError("", "Invalid code.");
-                    return View(model);
-            }
-        }
+//            // The following code protects for brute force attacks against the two factor codes. 
+//            // If a user enters incorrect codes for a specified amount of time then the user account 
+//            // will be locked out for a specified amount of time. 
+//            // You can configure the account lockout settings in IdentityConfig
+//            var result = await SignInManager.TwoFactorSignInAsync(model.Provider, model.Code, isPersistent:  model.RememberMe, rememberBrowser: model.RememberBrowser);
+//            switch (result)
+//            {
+//                case SignInStatus.Success:
+//                    return RedirectToLocal(model.ReturnUrl);
+//                case SignInStatus.LockedOut:
+//                    return View("Lockout");
+//                case SignInStatus.Failure:
+//                default:
+//                    ModelState.AddModelError("", "Invalid code.");
+//                    return View(model);
+//            }
+//        }
 
-        //
-        // GET: /Account/Register
-        [AllowAnonymous]
-        public Microsoft.AspNetCore.Mvc.ActionResult Register(string returnUrl)
-        {
-            ViewBag.ReturnUrl = returnUrl;
-            return View();
-        }
+//        //
+//        // GET: /Account/Register
+//        [AllowAnonymous]
+//        public Microsoft.AspNetCore.Mvc.ActionResult Register(string returnUrl)
+//        {
+//            ViewBag.ReturnUrl = returnUrl;
+//            return View();
+//        }
 
-        //
-        // POST: /Account/Register
-        [HttpPost]
-        [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> Register(RegisterViewModel model, string returnUrl)
-        {
-            if (ModelState.IsValid)
-            {
-                var user = new ApplicationUser { UserName = model.Email, Email = model.Email };
-                var result = await UserManager.CreateAsync(user, model.Password);
-                if (result.Succeeded)
-                {
-                    await SignInManager.SignInAsync(user, isPersistent:false, rememberBrowser:false);
+//        //
+//        // POST: /Account/Register
+//        [HttpPost]
+//        [AllowAnonymous]
+//        [ValidateAntiForgeryToken]
+//        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> Register(RegisterViewModel model, string returnUrl)
+//        {
+//            if (ModelState.IsValid)
+//            {
+//                var user = new ApplicationUser { UserName = model.Email, Email = model.Email };
+//                var result = await UserManager.CreateAsync(user, model.Password);
+//                if (result.Succeeded)
+//                {
+//                    await SignInManager.SignInAsync(user, isPersistent:false, rememberBrowser:false);
                     
-                    // For more information on how to enable account confirmation and password reset please visit http://go.microsoft.com/fwlink/?LinkID=320771
-                    // Send an email with this link
-                    if (_useSendGrid)
-                    {
-                        string code = await UserManager.GenerateEmailConfirmationTokenAsync(user.Id);
-                        var callbackUrl = Url.Action("ConfirmEmail", "Account",
-                            new {userId = user.Id, code = code},
-                            protocol: Request.Url.Scheme);
-                        await
-                            UserManager.SendEmailAsync(user.Id, "Confirm your email",
-                                "Please confirm your email address by clicking <a href=\"" + callbackUrl + "\">here</a>");
-                    }
-                    if (!string.IsNullOrWhiteSpace(returnUrl))
-                        return RedirectToLocal(returnUrl);
-                    return RedirectToAction("Start", "Home");
-                }
-                AddErrors(result);
-            }
+//                    // For more information on how to enable account confirmation and password reset please visit http://go.microsoft.com/fwlink/?LinkID=320771
+//                    // Send an email with this link
+//                    if (_useSendGrid)
+//                    {
+//                        string code = await UserManager.GenerateEmailConfirmationTokenAsync(user.Id);
+//                        var callbackUrl = Url.Action("ConfirmEmail", "Account",
+//                            new {userId = user.Id, code = code},
+//                            protocol: Request.Url.Scheme);
+//                        await
+//                            UserManager.SendEmailAsync(user.Id, "Confirm your email",
+//                                "Please confirm your email address by clicking <a href=\"" + callbackUrl + "\">here</a>");
+//                    }
+//                    if (!string.IsNullOrWhiteSpace(returnUrl))
+//                        return RedirectToLocal(returnUrl);
+//                    return RedirectToAction("Start", "Home");
+//                }
+//                AddErrors(result);
+//            }
 
-            // If we got this far, something failed, redisplay form
-            return View(model);
-        }
+//            // If we got this far, something failed, redisplay form
+//            return View(model);
+//        }
 
-        //
-        // GET: /Account/ConfirmEmail
-        [AllowAnonymous]
-        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> ConfirmEmail(string userId, string code)
-        {
-            if (userId == null || code == null)
-            {
-                return View("Error");
-            }
-            var user = await UserManager.GetUserAsync(User);
-            var result = await UserManager.ConfirmEmailAsync(user, code);
-            if (result.Succeeded)
-            {
-                // Add user to User role once email is confirmed
-                await UserManager.AddToRoleAsync(user, "User");
-            }
-            return View(result.Succeeded ? "ConfirmEmail" : "Error");
-        }
+//        //
+//        // GET: /Account/ConfirmEmail
+//        [AllowAnonymous]
+//        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> ConfirmEmail(string userId, string code)
+//        {
+//            if (userId == null || code == null)
+//            {
+//                return View("Error");
+//            }
+//            var user = await UserManager.GetUserAsync(User);
+//            var result = await UserManager.ConfirmEmailAsync(user, code);
+//            if (result.Succeeded)
+//            {
+//                // Add user to User role once email is confirmed
+//                await UserManager.AddToRoleAsync(user, "User");
+//            }
+//            return View(result.Succeeded ? "ConfirmEmail" : "Error");
+//        }
 
-        //
-        // GET: /Account/ForgotPassword
-        [AllowAnonymous]
-        public Microsoft.AspNetCore.Mvc.ActionResult ForgotPassword()
-        {
-            return View();
-        }
+//        //
+//        // GET: /Account/ForgotPassword
+//        [AllowAnonymous]
+//        public Microsoft.AspNetCore.Mvc.ActionResult ForgotPassword()
+//        {
+//            return View();
+//        }
 
-        //
-        // POST: /Account/ForgotPassword
-        [HttpPost]
-        [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> ForgotPassword(ForgotPasswordViewModel model)
-        {
-            if (ModelState.IsValid)
-            {
-                var user = await UserManager.FindByNameAsync(model.Email);
-                if (user == null || !(await UserManager.IsEmailConfirmedAsync(user)))
-                {
-                    //Notify unknown user that password reset attempt was made
-                    if (_useSendGrid)
-                    {
-                        await UserManager.EmailService.SendAsync(
-                            new IdentityMessage
-                            {
-                                Destination = model.Email,
-                                Subject = "Reset password",
-                                Body = string.Format(
-                                    @"Hi,<br/>
-<br/>
-Someone tried to reset the password for email {0} on our site.<br/>
-<br/>
-Unfortunately we can’t find any user with that address in our database. If you’d like to register a new account please use the link below:<br/>
-{1}
-",
-                                    model.Email,
-                                    Url.Action(
-                                        protocol: Request.Url.Scheme,
-                                        actionName: "Register",
-                                        controllerName: "Account",
-                                        routeValues: new {}
-                                        )
-                                    )
-                            }
-                            );
-                    }
+//        //
+//        // POST: /Account/ForgotPassword
+//        [HttpPost]
+//        [AllowAnonymous]
+//        [ValidateAntiForgeryToken]
+//        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> ForgotPassword(ForgotPasswordViewModel model)
+//        {
+//            if (ModelState.IsValid)
+//            {
+//                var user = await UserManager.FindByNameAsync(model.Email);
+//                if (user == null || !(await UserManager.IsEmailConfirmedAsync(user)))
+//                {
+//                    //Notify unknown user that password reset attempt was made
+//                    if (_useSendGrid)
+//                    {
+//                        await UserManager.EmailService.SendAsync(
+//                            new IdentityMessage
+//                            {
+//                                Destination = model.Email,
+//                                Subject = "Reset password",
+//                                Body = string.Format(
+//                                    @"Hi,<br/>
+//<br/>
+//Someone tried to reset the password for email {0} on our site.<br/>
+//<br/>
+//Unfortunately we can’t find any user with that address in our database. If you’d like to register a new account please use the link below:<br/>
+//{1}
+//",
+//                                    model.Email,
+//                                    Url.Action(
+//                                        protocol: Request.Url.Scheme,
+//                                        actionName: "Register",
+//                                        controllerName: "Account",
+//                                        routeValues: new {}
+//                                        )
+//                                    )
+//                            }
+//                            );
+//                    }
 
-                    // Don't reveal that the user does not exist or is not confirmed
-                    return View("ForgotPasswordConfirmation");
-                }
+//                    // Don't reveal that the user does not exist or is not confirmed
+//                    return View("ForgotPasswordConfirmation");
+//                }
 
-                // For more information on how to enable account confirmation and password reset please visit http://go.microsoft.com/fwlink/?LinkID=320771
-                // Send an email with this link
-                if (_useSendGrid)
-                {
-                    string code = await UserManager.GeneratePasswordResetTokenAsync(user.Id);
-                    var callbackUrl = Url.Action("ResetPassword", "Account",
-                        new {userId = user.Id, code = code},
-                        protocol: Request.Url.Scheme);
-                    await
-                        UserManager.SendEmailAsync(user.Id, "Reset Password",
-                            "Please reset your password by clicking <a href=\"" + callbackUrl + "\">here</a>");
-                }
+//                // For more information on how to enable account confirmation and password reset please visit http://go.microsoft.com/fwlink/?LinkID=320771
+//                // Send an email with this link
+//                if (_useSendGrid)
+//                {
+//                    string code = await UserManager.GeneratePasswordResetTokenAsync(user.Id);
+//                    var callbackUrl = Url.Action("ResetPassword", "Account",
+//                        new {userId = user.Id, code = code},
+//                        protocol: Request.Url.Scheme);
+//                    await
+//                        UserManager.SendEmailAsync(user.Id, "Reset Password",
+//                            "Please reset your password by clicking <a href=\"" + callbackUrl + "\">here</a>");
+//                }
 
-                return RedirectToAction("ForgotPasswordConfirmation", "Account");
-            }
+//                return RedirectToAction("ForgotPasswordConfirmation", "Account");
+//            }
 
-            // If we got this far, something failed, redisplay form
-            return View(model);
-        }
+//            // If we got this far, something failed, redisplay form
+//            return View(model);
+//        }
 
-        //
-        // GET: /Account/ForgotPasswordConfirmation
-        [AllowAnonymous]
-        public Microsoft.AspNetCore.Mvc.ActionResult ForgotPasswordConfirmation()
-        {
-            return View();
-        }
+//        //
+//        // GET: /Account/ForgotPasswordConfirmation
+//        [AllowAnonymous]
+//        public Microsoft.AspNetCore.Mvc.ActionResult ForgotPasswordConfirmation()
+//        {
+//            return View();
+//        }
 
-        //
-        // GET: /Account/ResetPassword
-        [AllowAnonymous]
-        public Microsoft.AspNetCore.Mvc.ActionResult ResetPassword(string code)
-        {
-            return code == null ? View("Error") : View();
-        }
+//        //
+//        // GET: /Account/ResetPassword
+//        [AllowAnonymous]
+//        public Microsoft.AspNetCore.Mvc.ActionResult ResetPassword(string code)
+//        {
+//            return code == null ? View("Error") : View();
+//        }
 
-        //
-        // POST: /Account/ResetPassword
-        [HttpPost]
-        [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> ResetPassword(ResetPasswordViewModel model)
-        {
-            if (!ModelState.IsValid)
-            {
-                return View(model);
-            }
-            var user = await UserManager.FindByNameAsync(model.Email);
-            if (user == null)
-            {
-                // Don't reveal that the user does not exist
-                return RedirectToAction("ResetPasswordConfirmation", "Account");
-            }
-            var result = await UserManager.ResetPasswordAsync(user, model.Code, model.Password);
-            if (result.Succeeded)
-            {
-                return RedirectToAction("ResetPasswordConfirmation", "Account");
-            }
-            AddErrors(result);
-            return View();
-        }
+//        //
+//        // POST: /Account/ResetPassword
+//        [HttpPost]
+//        [AllowAnonymous]
+//        [ValidateAntiForgeryToken]
+//        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> ResetPassword(ResetPasswordViewModel model)
+//        {
+//            if (!ModelState.IsValid)
+//            {
+//                return View(model);
+//            }
+//            var user = await UserManager.FindByNameAsync(model.Email);
+//            if (user == null)
+//            {
+//                // Don't reveal that the user does not exist
+//                return RedirectToAction("ResetPasswordConfirmation", "Account");
+//            }
+//            var result = await UserManager.ResetPasswordAsync(user, model.Code, model.Password);
+//            if (result.Succeeded)
+//            {
+//                return RedirectToAction("ResetPasswordConfirmation", "Account");
+//            }
+//            AddErrors(result);
+//            return View();
+//        }
 
-        //
-        // GET: /Account/ResetPasswordConfirmation
-        [AllowAnonymous]
-        public Microsoft.AspNetCore.Mvc.ActionResult ResetPasswordConfirmation()
-        {
-            return View();
-        }
+//        //
+//        // GET: /Account/ResetPasswordConfirmation
+//        [AllowAnonymous]
+//        public Microsoft.AspNetCore.Mvc.ActionResult ResetPasswordConfirmation()
+//        {
+//            return View();
+//        }
 
-        //
-        // POST: /Account/ExternalLogin
-        [HttpPost]
-        [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public Microsoft.AspNetCore.Mvc.ActionResult ExternalLogin(string provider, string returnUrl)
-        {
-            // Request a redirect to the external login provider
-            return new ChallengeResult(provider, Url.Action("ExternalLoginCallback", "Account", new { ReturnUrl = returnUrl }));
-        }
+//        //
+//        // POST: /Account/ExternalLogin
+//        [HttpPost]
+//        [AllowAnonymous]
+//        [ValidateAntiForgeryToken]
+//        public Microsoft.AspNetCore.Mvc.ActionResult ExternalLogin(string provider, string returnUrl)
+//        {
+//            // Request a redirect to the external login provider
+//            return new ChallengeResult(provider, Url.Action("ExternalLoginCallback", "Account", new { ReturnUrl = returnUrl }));
+//        }
 
-        //
-        // GET: /Account/SendCode
-        [AllowAnonymous]
-        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> SendCode(string returnUrl, bool rememberMe)
-        {
-            var userId = await SignInManager.GetVerifiedUserIdAsync();
-            if (userId == null)
-            {
-                return View("Error");
-            }
-            var userFactors = await UserManager.GetValidTwoFactorProvidersAsync(userId);
-            var factorOptions = userFactors.Select(purpose => new SelectListItem { Text = purpose, Value = purpose }).ToList();
-            return View(new SendCodeViewModel { Providers = factorOptions, ReturnUrl = returnUrl, RememberMe = rememberMe });
-        }
+//        //
+//        // GET: /Account/SendCode
+//        [AllowAnonymous]
+//        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> SendCode(string returnUrl, bool rememberMe)
+//        {
+//            var userId = await SignInManager.GetVerifiedUserIdAsync();
+//            if (userId == null)
+//            {
+//                return View("Error");
+//            }
+//            var userFactors = await UserManager.GetValidTwoFactorProvidersAsync(userId);
+//            var factorOptions = userFactors.Select(purpose => new SelectListItem { Text = purpose, Value = purpose }).ToList();
+//            return View(new SendCodeViewModel { Providers = factorOptions, ReturnUrl = returnUrl, RememberMe = rememberMe });
+//        }
 
-        //
-        // POST: /Account/SendCode
-        [HttpPost]
-        [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> SendCode(SendCodeViewModel model)
-        {
-            if (!ModelState.IsValid)
-            {
-                return View();
-            }
+//        //
+//        // POST: /Account/SendCode
+//        [HttpPost]
+//        [AllowAnonymous]
+//        [ValidateAntiForgeryToken]
+//        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> SendCode(SendCodeViewModel model)
+//        {
+//            if (!ModelState.IsValid)
+//            {
+//                return View();
+//            }
 
-            // Generate the token and send it
-            if (!await SignInManager.SendTwoFactorCodeAsync(model.SelectedProvider))
-            {
-                return View("Error");
-            }
-            return RedirectToAction("VerifyCode", new { Provider = model.SelectedProvider, ReturnUrl = model.ReturnUrl, RememberMe = model.RememberMe });
-        }
+//            // Generate the token and send it
+//            if (!await SignInManager.SendTwoFactorCodeAsync(model.SelectedProvider))
+//            {
+//                return View("Error");
+//            }
+//            return RedirectToAction("VerifyCode", new { Provider = model.SelectedProvider, ReturnUrl = model.ReturnUrl, RememberMe = model.RememberMe });
+//        }
 
-        //
-        // GET: /Account/ExternalLoginCallback
-        [AllowAnonymous]
-        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> ExternalLoginCallback(string returnUrl)
-        {
-            var loginInfo = await AuthenticationManager.GetExternalLoginInfoAsync();
-            if (loginInfo == null)
-            {
-                return RedirectToAction("Login");
-            }
+//        //
+//        // GET: /Account/ExternalLoginCallback
+//        [AllowAnonymous]
+//        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> ExternalLoginCallback(string returnUrl)
+//        {
+//            var loginInfo = await AuthenticationManager.GetExternalLoginInfoAsync();
+//            if (loginInfo == null)
+//            {
+//                return RedirectToAction("Login");
+//            }
 
-            // Sign in the user with this external login provider if the user already has a login
-            var result = await SignInManager.ExternalSignInAsync(loginInfo, isPersistent: false);
-            switch (result)
-            {
-                case SignInStatus.Success:
-                    return RedirectToLocal(returnUrl);
-                case SignInStatus.LockedOut:
-                    return View("Lockout");
-                case SignInStatus.RequiresVerification:
-                    return RedirectToAction("SendCode", new { ReturnUrl = returnUrl, RememberMe = false });
-                case SignInStatus.Failure:
-                default:
-                    // If the user does not have an account, then prompt the user to create an account
-                    ViewBag.ReturnUrl = returnUrl;
-                    ViewBag.LoginProvider = loginInfo.Login.LoginProvider;
-                    return View("ExternalLoginConfirmation", new ExternalLoginConfirmationViewModel { Email = loginInfo.Email });
-            }
-        }
+//            // Sign in the user with this external login provider if the user already has a login
+//            var result = await SignInManager.ExternalSignInAsync(loginInfo, isPersistent: false);
+//            switch (result)
+//            {
+//                case SignInStatus.Success:
+//                    return RedirectToLocal(returnUrl);
+//                case SignInStatus.LockedOut:
+//                    return View("Lockout");
+//                case SignInStatus.RequiresVerification:
+//                    return RedirectToAction("SendCode", new { ReturnUrl = returnUrl, RememberMe = false });
+//                case SignInStatus.Failure:
+//                default:
+//                    // If the user does not have an account, then prompt the user to create an account
+//                    ViewBag.ReturnUrl = returnUrl;
+//                    ViewBag.LoginProvider = loginInfo.Login.LoginProvider;
+//                    return View("ExternalLoginConfirmation", new ExternalLoginConfirmationViewModel { Email = loginInfo.Email });
+//            }
+//        }
 
-        //
-        // POST: /Account/ExternalLoginConfirmation
-        [HttpPost]
-        [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> ExternalLoginConfirmation(ExternalLoginConfirmationViewModel model, string returnUrl)
-        {
-            if (User.Identity.IsAuthenticated)
-            {
-                return RedirectToAction("Index", "Manage");
-            }
+//        //
+//        // POST: /Account/ExternalLoginConfirmation
+//        [HttpPost]
+//        [AllowAnonymous]
+//        [ValidateAntiForgeryToken]
+//        public async Task<Microsoft.AspNetCore.Mvc.ActionResult> ExternalLoginConfirmation(ExternalLoginConfirmationViewModel model, string returnUrl)
+//        {
+//            if (User.Identity.IsAuthenticated)
+//            {
+//                return RedirectToAction("Index", "Manage");
+//            }
 
-            if (ModelState.IsValid)
-            {
-                // Get the information about the user from the external login provider
-                var info = await AuthenticationManager.GetExternalLoginInfoAsync();
-                if (info == null)
-                {
-                    return View("ExternalLoginFailure");
-                }
-                var user = new ApplicationUser { UserName = model.Email, Email = model.Email };
-                var result = await UserManager.CreateAsync(user);
-                if (result.Succeeded)
-                {
-                    result = await UserManager.AddLoginAsync(user.Id, info.Login);
-                    if (result.Succeeded)
-                    {
-                        await SignInManager.SignInAsync(user, isPersistent: false, rememberBrowser: false);
+//            if (ModelState.IsValid)
+//            {
+//                // Get the information about the user from the external login provider
+//                var info = await AuthenticationManager.GetExternalLoginInfoAsync();
+//                if (info == null)
+//                {
+//                    return View("ExternalLoginFailure");
+//                }
+//                var user = new ApplicationUser { UserName = model.Email, Email = model.Email };
+//                var result = await UserManager.CreateAsync(user);
+//                if (result.Succeeded)
+//                {
+//                    result = await UserManager.AddLoginAsync(user.Id, info.Login);
+//                    if (result.Succeeded)
+//                    {
+//                        await SignInManager.SignInAsync(user, isPersistent: false, rememberBrowser: false);
 
-                        if (_useSendGrid)
-                        {
-                            string code = await UserManager.GenerateEmailConfirmationTokenAsync(user);
-                            var callbackUrl = Url.Action("ConfirmEmail", "Account",
-                                new { userId = user.Id, code = code },
-                                protocol: Request.Url.Scheme);
-                            await
-                                UserManager.SendEmailAsync(user.Id, "Confirm your email",
-                                    "Please confirm your email address by clicking <a href=\"" + callbackUrl + "\">here</a>");
-                        }
+//                        if (_useSendGrid)
+//                        {
+//                            string code = await UserManager.GenerateEmailConfirmationTokenAsync(user);
+//                            var callbackUrl = Url.Action("ConfirmEmail", "Account",
+//                                new { userId = user.Id, code = code },
+//                                protocol: Request.Url.Scheme);
+//                            await
+//                                UserManager.SendEmailAsync(user.Id, "Confirm your email",
+//                                    "Please confirm your email address by clicking <a href=\"" + callbackUrl + "\">here</a>");
+//                        }
 
-                        return RedirectToLocal(returnUrl);
-                    }
-                }
-                AddErrors(result);
-            }
+//                        return RedirectToLocal(returnUrl);
+//                    }
+//                }
+//                AddErrors(result);
+//            }
 
-            ViewBag.ReturnUrl = returnUrl;
-            return View(model);
-        }
+//            ViewBag.ReturnUrl = returnUrl;
+//            return View(model);
+//        }
 
-        //
-        // POST: /Account/LogOff
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public Microsoft.AspNetCore.Mvc.ActionResult LogOff()
-        {
-            AuthenticationManager.SignOut(DefaultAuthenticationTypes.ApplicationCookie);
-            return RedirectToAction("Start", "Home");
-        }
+//        //
+//        // POST: /Account/LogOff
+//        [HttpPost]
+//        [ValidateAntiForgeryToken]
+//        public Microsoft.AspNetCore.Mvc.ActionResult LogOff()
+//        {
+//            AuthenticationManager.SignOut(DefaultAuthenticationTypes.ApplicationCookie);
+//            return RedirectToAction("Start", "Home");
+//        }
 
-        //
-        // GET: /Account/ExternalLoginFailure
-        [AllowAnonymous]
-        public Microsoft.AspNetCore.Mvc.ActionResult ExternalLoginFailure()
-        {
-            return View();
-        }
+//        //
+//        // GET: /Account/ExternalLoginFailure
+//        [AllowAnonymous]
+//        public Microsoft.AspNetCore.Mvc.ActionResult ExternalLoginFailure()
+//        {
+//            return View();
+//        }
 
-        protected override void Dispose(bool disposing)
-        {
-            if (disposing)
-            {
-                if (_userManager != null)
-                {
-                    _userManager.Dispose();
-                    _userManager = null;
-                }
+//        protected override void Dispose(bool disposing)
+//        {
+//            if (disposing)
+//            {
+//                if (_userManager != null)
+//                {
+//                    _userManager.Dispose();
+//                    _userManager = null;
+//                }
 
-                if (_signInManager != null)
-                {
-                    _signInManager.Dispose();
-                    _signInManager = null;
-                }
-            }
+//                if (_signInManager != null)
+//                {
+//                    _signInManager.Dispose();
+//                    _signInManager = null;
+//                }
+//            }
 
-            base.Dispose(disposing);
-        }
+//            base.Dispose(disposing);
+//        }
 
-        #region Helpers
-        // Used for XSRF protection when adding external logins
-        // private const string XsrfKey = "XsrfId";
-        private const string XsrfKey = "feBGOWgHWz4tWyxbmBBN";
+//        #region Helpers
+//        // Used for XSRF protection when adding external logins
+//        // private const string XsrfKey = "XsrfId";
+//        private const string XsrfKey = "feBGOWgHWz4tWyxbmBBN";
 
-        private IAuthenticationManager AuthenticationManager
-        {
-            get
-            {
-                return HttpContext.GetOwinContext().Authentication;
-            }
-        }
+//        private IAuthenticationManager AuthenticationManager
+//        {
+//            get
+//            {
+//                return HttpContext.GetOwinContext().Authentication;
+//            }
+//        }
 
-        private void AddErrors(IdentityResult result)
-        {
-            foreach (var error in result.Errors)
-            {
-                ModelState.AddModelError("", error);
-            }
-        }
+//        private void AddErrors(IdentityResult result)
+//        {
+//            foreach (var error in result.Errors)
+//            {
+//                ModelState.AddModelError("", error);
+//            }
+//        }
 
-        private Microsoft.AspNetCore.Mvc.ActionResult RedirectToLocal(string returnUrl)
-        {
-            if (Url.IsLocalUrl(returnUrl))
-            {
-                return Redirect(returnUrl);
-            }
-            return RedirectToAction("Start", "Home");
-        }
+//        private Microsoft.AspNetCore.Mvc.ActionResult RedirectToLocal(string returnUrl)
+//        {
+//            if (Url.IsLocalUrl(returnUrl))
+//            {
+//                return Redirect(returnUrl);
+//            }
+//            return RedirectToAction("Start", "Home");
+//        }
 
-        internal class ChallengeResult : Microsoft.AspNetCore.Mvc.UnauthorizedResult
-        {
-            public ChallengeResult(string provider, string redirectUri)
-                : this(provider, redirectUri, null)
-            {
-            }
+//        internal class ChallengeResult : Microsoft.AspNetCore.Mvc.UnauthorizedResult
+//        {
+//            public ChallengeResult(string provider, string redirectUri)
+//                : this(provider, redirectUri, null)
+//            {
+//            }
 
-            public ChallengeResult(string provider, string redirectUri, string userId)
-            {
-                LoginProvider = provider;
-                RedirectUri = redirectUri;
-                UserId = userId;
-            }
+//            public ChallengeResult(string provider, string redirectUri, string userId)
+//            {
+//                LoginProvider = provider;
+//                RedirectUri = redirectUri;
+//                UserId = userId;
+//            }
 
-            public string LoginProvider { get; set; }
-            public string RedirectUri { get; set; }
-            public string UserId { get; set; }
+//            public string LoginProvider { get; set; }
+//            public string RedirectUri { get; set; }
+//            public string UserId { get; set; }
 
-            public override void ExecuteResult(ControllerContext context)
-            {
-                var properties = new AuthenticationProperties { RedirectUri = RedirectUri };
-                if (UserId != null)
-                {
-                    properties.Dictionary[XsrfKey] = UserId;
-                }
-                context.HttpContext.GetOwinContext().Authentication.Challenge(properties, LoginProvider);
-            }
-        }
-        #endregion
-    }
-}
+//            public override void ExecuteResult(ControllerContext context)
+//            {
+//                var properties = new AuthenticationProperties { RedirectUri = RedirectUri };
+//                if (UserId != null)
+//                {
+//                    properties.Dictionary[XsrfKey] = UserId;
+//                }
+//                context.HttpContext.GetOwinContext().Authentication.Challenge(properties, LoginProvider);
+//            }
+//        }
+//        #endregion
+//    }
+//}
