@@ -1,12 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Data.Entity;
+using Microsoft.EntityFrameworkCore;
 using System.Globalization;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Swetugg.Web.Models;
 using Swetugg.Web.Services;
+using Microsoft.Extensions.Caching.Memory;
 
 namespace Swetugg.Web.Controllers
 {
@@ -39,10 +40,10 @@ namespace Swetugg.Web.Controllers
         private string conferenceSlug;
         private Conference conference;
 
-        public ScheduleApiController(ApplicationDbContext dbContext)
+        public ScheduleApiController(ApplicationDbContext dbContext, IMemoryCache memoryCache)
         {
             this.dbContext = dbContext;
-            this.conferenceService = new CachedConferenceService(new ConferenceService(dbContext));
+            this.conferenceService = new CachedConferenceService(new ConferenceService(dbContext), memoryCache);
 
         }
 
@@ -135,7 +136,7 @@ namespace Swetugg.Web.Controllers
             {
                 if (conference != null)
                     return conference;
-                return conference = dbContext.Conferences.SingleAsync(c => c.Slug == ConferenceSlug);
+                return conference = dbContext.Conferences.Single(c => c.Slug == ConferenceSlug);
             }
 
         }
