@@ -1,15 +1,15 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿//using Microsoft.AspNetCore.Mvc;
 
-namespace Swetugg.Web.Helpers
-{
-    public static class MarkdownHelpers
-    {
-        public static string FromMarkdown(this HtmlHelper helper, string markdown)
-        {
-            var md = new MarkdownDeep.Markdown();
-            md.SafeMode = true;
-            md.ExtraMode = true;
-            return md.Transform(markdown);
-        }
-    }
-}
+//namespace Swetugg.Web.Helpers
+//{
+//    public static class MarkdownHelpers
+//    {
+//        public static string FromMarkdown(this HtmlHelper helper, string markdown)
+//        {
+//            var md = new MarkdownDeep.Markdown();
+//            md.SafeMode = true;
+//            md.ExtraMode = true;
+//            return md.Transform(markdown);
+//        }
+//    }
+//}
