@@ -1,4 +1,4 @@
-﻿@model HandleErrorInfo
+﻿@*@model HandleErrorInfo*@
 
 @{
     ViewBag.Title = "Error";
@@ -7,6 +7,8 @@
 <h1 class="text-danger">Error.</h1>
 <h2 class="text-danger">An error occurred while processing your request.</h2>
 
+@*
+// TODO: 
 @if (Model.Exception != null)
 {
     <!--
@@ -18,4 +20,4 @@
     @Model.Exception.StackTrace
 
     -->
-}
+}*@
