@@ -1,6 +1,4 @@
-﻿@model System.Web.Mvc.HandleErrorInfo
-
-@{
+﻿@{
     ViewBag.Title = "Locked Out";
 }
 
